requirements.txt
alembic==1.15.1
aniso8601==10.0.0
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
Flask==3.1.0
Flask-Caching==2.3.1
flask-cors==5.0.1
Flask-Limiter==3.10.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SeaSurf==2.0.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
limits==4.1
logzero==1.7.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
Pygments==2.19.1
PyMySQL==1.1.1
python-dotenv==1.0.1
python-engineio==4.12.0
python-socketio==5.13.0
pytz==2025.1
requests==2.32.3
rich==13.9.4
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
PyJWT==2.9.0
firebase-admin==6.4.0
gevent==24.2.1
gevent-websocket==0.10.1
qrcode==8.0
Pillow==11.1.0
